Share and Enjoy! (Share common TypeExtensions code)
