+++ src/Microsoft.Data.Entity.InMemory/Microsoft.Data.Entity.InMemory.csproj
@@ -74,7 +74,7 @@
     <Compile Include="Utilities\Check.cs" />
     <Compile Include="..\Shared\CodeAnnotations.cs" />
     <Compile Include="..\Shared\StringBuilderExtensions.cs" />
-    <Compile Include="Utilities\TypeExtensions.cs" />
+    <Compile Include="..\Shared\SharedTypeExtensions.cs" />
     <EmbeddedResource Include="Properties\Strings.resx">
       <LogicalName>Microsoft.Data.Entity.InMemory.Strings.resources</LogicalName>
     </EmbeddedResource>
